Fix: Add v2 to the url of the new endpoint
