+++ api/src/main/java/ca/bc/gov/educ/api/student/profile/email/controller/v1/GetMyPenEmailController.java
@@ -1,11 +1,11 @@
-package ca.bc.gov.educ.api.student.profile.email.controller;
+package ca.bc.gov.educ.api.student.profile.email.controller.v1;
 
 
-import ca.bc.gov.educ.api.student.profile.email.endpoint.GetMyPenEmailEndpoint;
+import ca.bc.gov.educ.api.student.profile.email.endpoint.v1.GetMyPenEmailEndpoint;
 import ca.bc.gov.educ.api.student.profile.email.exception.InvalidPayloadException;
 import ca.bc.gov.educ.api.student.profile.email.exception.errors.ApiError;
 import ca.bc.gov.educ.api.student.profile.email.service.GMPEmailService;
-import ca.bc.gov.educ.api.student.profile.email.struct.gmpump.*;
+import ca.bc.gov.educ.api.student.profile.email.struct.v1.gmpump.*;
 import ca.bc.gov.educ.api.student.profile.email.validator.EmailValidator;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
