+++ api/src/main/java/ca/bc/gov/educ/api/student/profile/email/service/EmailNotificationService.java
@@ -1,6 +1,7 @@
 package ca.bc.gov.educ.api.student.profile.email.service;
 
-import ca.bc.gov.educ.api.student.profile.email.struct.EmailNotificationEntity;
+import ca.bc.gov.educ.api.student.profile.email.exception.StudentEmailRuntimeException;
+import ca.bc.gov.educ.api.student.profile.email.struct.v2.EmailNotificationEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -37,7 +38,11 @@ public class EmailNotificationService {
     final var ctx = new Context();
     emailNotificationEntity.getEmailFields().forEach(ctx::setVariable);
 
-    final var body = this.templateEngine.process(this.templateConfig.get(emailNotificationEntity.getTemplateName()), ctx);   //emailNotificationEntity.getTemplateName() + ".html"
+    if(!this.templateConfig.containsKey(emailNotificationEntity.getTemplateName())) {
+      throw new StudentEmailRuntimeException("Not found email template for template name :: " +  emailNotificationEntity.getTemplateName());
+    }
+
+    final var body = this.templateEngine.process(this.templateConfig.get(emailNotificationEntity.getTemplateName()), ctx);
 
     this.getChesEmailService().sendEmail(emailNotificationEntity.getFromEmail(), emailNotificationEntity.getToEmail(), body, emailNotificationEntity.getSubject());
     log.debug("Email sent successfully");
