--- api/src/main/java/ca/bc/gov/educ/api/student/profile/email/struct/penrequestbatch/PendingRecords.java
@@ -1,7 +1,0 @@
-package ca.bc.gov.educ.api.student.profile.email.struct.penrequestbatch;
-
-public enum PendingRecords {
-  ALL,
-  SOME,
-  NONE
-}
