+++ api/src/main/java/ca/bc/gov/educ/api/student/profile/email/struct/v1/gmpump/GMPRequestRejectedEmailEntity.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.educ.api.student.profile.email.struct.gmpump;
+package ca.bc.gov.educ.api.student.profile.email.struct.v1.gmpump;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
@@ -11,9 +11,9 @@ import javax.validation.constraints.NotNull;
 @Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
 @ToString(callSuper = true)
-public class UMPRequestCompleteEmailEntity extends BaseEmailEntity {
+public class GMPRequestRejectedEmailEntity extends BaseEmailEntity {
 
-  @NotNull(message = "First Name can not be null")
-  private String firstName;
+  @NotNull(message = "Rejection Reason can not be null")
+  private String rejectionReason;
 
 }
