+++ api/src/main/java/ca/bc/gov/educ/api/student/profile/email/struct/v1/gmpump/UMPRequestRejectedEmailEntity.java
@@ -1,17 +1,18 @@
-package ca.bc.gov.educ.api.student.profile.email.struct.gmpump;
+package ca.bc.gov.educ.api.student.profile.email.struct.v1.gmpump;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
+import lombok.*;
 
 import javax.validation.constraints.NotNull;
 
-@Getter
-@Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+@EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
-public class GMPRequestRejectedEmailEntity extends BaseEmailEntity {
+public class UMPRequestRejectedEmailEntity extends BaseEmailEntity {
 
   @NotNull(message = "Rejection Reason can not be null")
   private String rejectionReason;
