+++ api/src/main/java/ca/bc/gov/educ/api/student/profile/email/struct/v1/penrequestbatch/PendingRecords.java
@@ -1,0 +1,7 @@
+package ca.bc.gov.educ.api.student.profile.email.struct.v1.penrequestbatch;
+
+public enum PendingRecords {
+  ALL,
+  SOME,
+  NONE
+}
