+++ api/src/main/java/ca/bc/gov/educ/api/student/profile/email/struct/v2/EmailNotificationEntity.java
@@ -1,10 +1,11 @@
-package ca.bc.gov.educ.api.student.profile.email.struct;
+package ca.bc.gov.educ.api.student.profile.email.struct.v2;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.NotNull;
 import java.util.Map;
 
 @Data
@@ -12,11 +13,17 @@ import java.util.Map;
 @AllArgsConstructor
 @NoArgsConstructor
 public class EmailNotificationEntity {
+  @NotNull(message = "fromEmail can not be null.")
   private String fromEmail;
+  @NotNull(message = "toEmail can not be null.")
   private String toEmail;
+  @NotNull(message = "subject can not be null.")
   private String subject;
 
+  @NotNull(message = "templateName can not be null.")
   private String templateName;
+
+  @NotNull(message = "emailFields can not be null.")
   private Map<String, String> emailFields;
 }
