+++ api/src/test/java/ca/bc/gov/educ/api/student/profile/email/service/EventHandlerDelegatorServiceTest.java
@@ -4,12 +4,11 @@ import ca.bc.gov.educ.api.student.profile.email.constants.EventType;
 import ca.bc.gov.educ.api.student.profile.email.props.ApplicationProperties;
 import ca.bc.gov.educ.api.student.profile.email.repository.StudentProfileRequestEmailEventRepository;
 import ca.bc.gov.educ.api.student.profile.email.rest.RestUtils;
-import ca.bc.gov.educ.api.student.profile.email.struct.EmailNotificationEntity;
+import ca.bc.gov.educ.api.student.profile.email.struct.v2.EmailNotificationEntity;
 import ca.bc.gov.educ.api.student.profile.email.struct.Event;
-import ca.bc.gov.educ.api.student.profile.email.struct.gmpump.*;
-import ca.bc.gov.educ.api.student.profile.email.struct.macro.MacroEditNotificationEntity;
-import ca.bc.gov.educ.api.student.profile.email.struct.penrequestbatch.ArchivePenRequestBatchNotificationEntity;
-import ca.bc.gov.educ.api.student.profile.email.struct.penrequestbatch.PenRequestBatchSchoolErrorNotificationEntity;
+import ca.bc.gov.educ.api.student.profile.email.struct.v1.gmpump.*;
+import ca.bc.gov.educ.api.student.profile.email.struct.v1.macro.*;
+import ca.bc.gov.educ.api.student.profile.email.struct.v1.penrequestbatch.*;
 import ca.bc.gov.educ.api.student.profile.email.utils.JsonUtil;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.val;
